requirements.txt
pillow
requests
scipy
sounddevice
numpy<2
shazamio
librosa
inky
tflite-runtime
typing
pyyaml
spotipy
openai
astral>=2.2